+++ moveit_experimental/moveit_jog_arm/CMakeLists.txt
@@ -126,17 +126,8 @@ if(CATKIN_ENABLE_TESTING)
   find_package(rostest REQUIRED)
   find_package(ros_pytest REQUIRED)
 
-  # Python integration test of basic message reception/publication
-  add_rostest(test/launch/jog_arm_msg_reception_test.test)
-
   # Python integration test that a halt message is published at a singularity
   add_rostest(test/launch/jog_arm_halt_msg_test.test)
-
-  # Python integration test of a service to enable drift in some dimensions
-  add_rostest(test/launch/jog_arm_drift_dimensions_service_test.test)
-
-  # Python integration test of velocity and acceleration limits
-  add_rostest(test/launch/jog_arm_vel_accel_limits_test.test)
 
   # jog_cpp_interface
   add_rostest_gtest(jog_cpp_interface_test
