+++ moveit_experimental/moveit_jog_arm/test/launch/jog_arm_drift_dimensions_service_test.test
@@ -10,6 +10,6 @@
     <rosparam command="load" file="$(find moveit_jog_arm)/test/arm_setup/config/initial_position.yaml" />
   </node>
 
-  <param name="test_module" value="$(find moveit_jog_arm)/test/test_drift_dimensions_service" />
+  <param name="test_module" value="$(find moveit_jog_arm)/test/python_tests/drift_dimensions_service" />
   <test test-name="jog_arm_drift_dimensions_service_test" pkg="ros_pytest" type="ros_pytest_runner" />
 </launch>
