+++ moveit_experimental/moveit_jog_arm/test/launch/jog_arm_halt_msg_test.test
@@ -10,8 +10,6 @@
 	  <rosparam command="load" file="$(find moveit_jog_arm)/test/arm_setup/config/singular_position.yaml" />
   </node>
 
-<!--
-  <param name="test_module" value="$(find moveit_jog_arm)/test/test_halt_msg" />
+  <param name="test_module" value="$(find moveit_jog_arm)/test/python_tests/halt_msg" />
   <test test-name="jog_arm_halt_msg_test" pkg="ros_pytest" type="ros_pytest_runner" />
--->
 </launch>
